CODES IN MATLAB
The following code helps us to evaluate or exercise newton's forward interpolation problem in MATLAB to find out the value of y=tan(x) when a range of x and y values are provided. 

Sample data tested on MATLAB:

Enter the total no. of data : 
5
Enter the coordinates of x : 
[25;50;75;100;125]

Enter the coordinates of y : 
[100;121;145;196;200]

Enter the value of x :
65
Table : 
   100    21     3    24    -98
   121    24    27  -74       0
   145    51   -47     0       0
   196      4      0      0      0
   200      0      0      0      0

The value of fn( 65 ) is:  131.3088